refactor: centralize root runtime sources
