+++ rustfs/src/app/object_usecase.rs
@@ -44,7 +44,7 @@ use super::{
     tagging::decode_tags,
     versioning_sys::BucketVersioningSys,
 };
-use crate::app::context::{
+use crate::app::runtime_sources::{
     AppContext, get_global_app_context, resolve_expiry_state_handle, resolve_notify_interface_for_context,
     resolve_object_store_handle_for_context,
 };
