+++ rustfs/src/startup_iam.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::app::context::AppContext;
+use crate::runtime_sources::AppContext;
 use crate::server::{ServiceStateManager, publish_ready_when_runtime_ready};
 use crate::storage::ECStore;
 use rustfs_common::{GlobalReadiness, SystemStage};
